# Avoid computing group indices when possible