This can make simple aggregation functions more than twice faster.

Some benchmarking shows no regression from current state even when a custom function (which doesn't benefit from optimization) is used:
```julia
using DataFrames, PooledArrays, BenchmarkTools
n = 100_000;
for k in (10, 10_000)
    global df = DataFrame(x=PooledArray(rand(1:k, n)), y=rand(n));
    println("groupby with $k groups")
    @btime groupby(df, :x);
    global gd = groupby(df, :x);

    println("sum with $k groups")
    @btime by(df, :x, :y => sum);
    @btime combine(gd2, :y => sum) setup=(gd2=deepcopy(gd));

    println("sum+length with $k groups")
    @btime by(df, :x, :y => sum, :y => length);
    @btime combine(gd2, :y => sum, :y => length) setup=(gd2=deepcopy(gd));

    println("x -> x[1] with $k groups")
    @btime by(df, :x, :y => (x -> x[1]));
    @btime combine(gd2, :y => (x -> x[1])) setup=(gd2=deepcopy(gd));
end

# Before
groupby with 10 groups
  463.561 μs (49 allocations: 1.53 MiB)
sum with 10 groups
  612.750 μs (188 allocations: 1.54 MiB)
  97.593 μs (139 allocations: 10.28 KiB)
sum+length with 10 groups
  599.455 μs (220 allocations: 1.54 MiB)
  106.351 μs (171 allocations: 12.14 KiB)
x -> x[1] with 10 groups
  702.722 μs (269 allocations: 2.30 MiB)
  153.622 μs (220 allocations: 793.52 KiB) # Indices have already been computed by first run
groupby with 10000 groups
  1.505 ms (52 allocations: 1.77 MiB)
sum with 10000 groups
  1.881 ms (210 allocations: 2.35 MiB)
  382.246 μs (158 allocations: 596.09 KiB)
sum+length with 10000 groups
  2.099 ms (253 allocations: 2.66 MiB)
  475.889 μs (201 allocations: 910.25 KiB)
x -> x[1] with 10000 groups
  4.099 ms (89172 allocations: 5.12 MiB)
  2.436 ms (89120 allocations: 3.35 MiB) # Indices have already been computed by first run

# After
groupby with 10 groups
  152.112 μs (44 allocations: 784.13 KiB)
sum with 10 groups
  271.348 μs (173 allocations: 794.02 KiB)
  92.858 μs (129 allocations: 9.89 KiB)
sum+length with 10 groups
  362.215 μs (205 allocations: 795.77 KiB)
  180.631 μs (161 allocations: 11.64 KiB) # Counts groups again every time
x -> x[1] with 10 groups
  689.666 μs (264 allocations: 2.30 MiB)
  152.599 μs (215 allocations: 793.34 KiB) # Indices have already been computed by first run
groupby with 10000 groups
  235.042 μs (45 allocations: 872.06 KiB)
sum with 10000 groups
  780.547 μs (193 allocations: 1.44 MiB)
  524.955 μs (148 allocations: 605.55 KiB)
sum+length with 10000 groups
  997.213 μs (235 allocations: 1.67 MiB)
  710.533 μs (190 allocations: 841.55 KiB) # Counts groups again every time
x -> x[1] with 10000 groups
  4.060 ms (89186 allocations: 5.12 MiB)
  2.540 ms (89133 allocations: 3.35 MiB) # Indices have already been computed by first run
```