@bkamins 2019-12-08
Wy do we have to mage `GroupedDataFrame` mutable for this? Because of `ngroups`?