@nalimilan 2019-12-08
No (it doesn't change), that's because of the nullable fields which need to be populated on the first call to `compute_indices!` to avoid recomputing them several times.