@bkamins 2019-12-08
Not sure what would be better, but we could make them empty like `Int[]`, and then check if they are empty by `isempty`.

The only corner case to check against would be if it is possible to have them empty in practice (e.g. a grouping of an empty data frame, or an empty `GroupedDataFrame` as a result of selection of zero entries - in the latter case it seems that `starts` and `ends` are empty but not `idx`)..

I have just checked - you can `groupby` an empty `DataFrame`:
```
julia> groupby(DataFrame(a=Int[], b=Int[]), :a)
GroupedDataFrame with 0 groups based on key: a

julia> by(DataFrame(a=Int[], b=Int[]), :a, nrow)
0×1 DataFrame
```

So I am not sure what would be best (maybe your design is cleaner and I guess there is no much harm in having `GroupedDataFrame` mutable - right?)