@nalimilan 2019-12-08
Yeah, we could check whether `gd.ngroups == length(gd.idx)`,  but using `nothing` is so much cleaner for this. Actually, it allowed me to catch very easily all places where `compute_indices!(gd)` was needed. Is there any clear drawback to using a mutable struct? Maybe we need to store the fields in variables to avoid slowing down some loops, but for users it shouldn't matter, right?