@bkamins 2019-12-08
> Maybe we need to store the fields in variables to avoid slowing down some loops, but for users it shouldn't matter, right?

This is what I would assume. It is just a reflex to avoid `mutable struct`, but in this case it is a heavy object anyway so it should be OK.

I will go through in detail through this PR once the CI start passing as it now fails (or you prefer me to check it now?).