@bkamins 2019-12-09
This is a complex PR. I understand you do two changes in it:
1. code that makes sure we do lazy evaluation of group indices (only when they are needed)
2. code that performs computations that do not require doing point 1. above if this is possible

The point 2. is rather clear (I assume that we handle `sort` and `skipmissing` in this case efficiently here also - right?).

However, I see that you did some changes to the logic of point 1. (e.g. regarding `sort` and `skipmissing` options) - can you please add some comment how we avoid this? (the code is "compressed", so I would prefer to make sure I correctly read its logic). Thank you