@nalimilan 2019-12-13
> However, I see that you did some changes to the logic of point 1. (e.g. regarding `sort` and `skipmissing` options) - can you please add some comment how we avoid this? (the code is "compressed", so I would prefer to make sure I correctly read its logic). Thank you

So the PR was indeed a bit messy in that regard. I've tried to make this cleaner by computing directly indices in `0:ngroups`, with 0 always referring to rows which are supposed to be skipped. That way the first group corresponds to index 1 everywhere (instead of subtracting 1 when `skipmissing=true`). I've also added comments which should make this more explicit, let me know what you think.