@bkamins 2019-12-14
Reading through this PR I have found the following problem:
```
julia> df1 = DataFrame(x=["a", "b", "c"], y=["1", "2", "3"])
3×2 DataFrame
│ Row │ x      │ y      │
│     │ String │ String │
├─────┼────────┼────────┤
│ 1   │ a      │ 1      │
│ 2   │ b      │ 2      │
│ 3   │ c      │ 3      │

julia> df2 = categorical(df1, :x)
3×2 DataFrame
│ Row │ x            │ y      │
│     │ Categorical… │ String │
├─────┼──────────────┼────────┤
│ 1   │ a            │ 1      │
│ 2   │ b            │ 2      │
│ 3   │ c            │ 3      │

julia> levels!(df2.x, ["b", "c", "a"])
3-element CategoricalArray{String,1,UInt32}:
 "a"
 "b"
 "c"

julia> by(df1, :x, identity)
3×2 DataFrame
│ Row │ x      │ y      │
│     │ String │ String │
├─────┼────────┼────────┤
│ 1   │ a      │ 1      │
│ 2   │ b      │ 2      │
│ 3   │ c      │ 3      │

julia> by(df2, :x, identity)
3×2 DataFrame
│ Row │ x            │ y      │
│     │ Categorical… │ String │
├─────┼──────────────┼────────┤
│ 1   │ b            │ 2      │
│ 2   │ c            │ 3      │
│ 3   │ a            │ 1      │
```
In other words - we always sort in `by`/`groupby` if we pass categorical grouping columns, which is inconsistent with what we promise (this should only happen if `sort=true` is passed).